+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
 ## Unreleased
-- fixed bug when target region wasn't cleared after check
+
+- fixed bug when target region wasn't cleared after check ([Trello](https://trello.com/c/gMwZw0C0/268-wdio5-cannot-read-property-offset-of-null-when-taking-window-screenshot-after-region))
 
 ## 5.9.17